initial page refacTOR
