+++ package.json
@@ -11,6 +11,7 @@
   "dependencies": {
     "@types/classnames": "^2.3.1",
     "@types/semver": "^7.5.3",
+    "@wagmi/core": "^1.4.6",
     "bitecs": "^0.3.40",
     "classnames": "^2.3.2",
     "dotenv": "^16.3.1",
@@ -22,7 +23,9 @@
     "react": "^18",
     "react-dom": "^18",
     "regenerator-runtime": "^0.14.0",
-    "semver": "^7.5.4"
+    "semver": "^7.5.4",
+    "viem": "^1.18.9",
+    "wagmi": "^1.4.6"
   },
   "devDependencies": {
     "@types/node": "^20",
